detect os
